Ship Drag Force

Compute the retarding force of water on the ship, based on a drag law \((1/2)*rho*Cs*A*vs^2\) where rho is water density taken to be 1024 (kg/m^3), Cs is drag coefficient stored in parms and A is area, also stored in parms, and vs` is the ship speed (m/s).

Usage

shipWaterForce(vs, parms)

Arguments

vs

Ship speed in m/s. (Consider using knot2mps() if you prefer to think of speeds in knots.)

parms

A named list holding model parameters, created by parameters().

Value

Water drag force (N).

See also

Author

Dan Kelley